requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
numpy==1.26.4
scikit-learn==1.4.1.post1
sentence-transformers==2.2.2
torch==2.1.0
faiss-cpu==1.7.4
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
apscheduler==3.10.4